chore: move car feature to domain package
