+++ src/main/java/racingcar/view/ResultView.java
@@ -1,6 +1,6 @@
 package racingcar.view;
 
-import racingcar.Car;
+import racingcar.domain.Car;
 
 import java.util.List;
 import java.util.StringJoiner;
